[finish#163278985] adds user details on registration with passing tests
